an inital attempt in UI
used pygame, others must install it too!
cards directory and its content are just a placeholder.
inteface.py file has some assumptions, which are explained in the
comments.
a click on any card at hand is tracked, at a "closed card" is not.
the function determining the responce for a click is not yet
implemented.